closes #167; changes to msysgit download page
Thanks @justmarkham for improving this; the old way was sort of archaic.
Investigating, I changed to link directly to the
msysgit (windows-specific) page, so that the three platforms are each
directed to something specifically for them (nobody gets directed to the
git-scm site and then has to choose again based on their platform).